Resolve rebase conflict
